+++ graphql-spqr-spring-boot-autoconfigure/src/main/java/io/leangen/graphql/spqr/spring/autoconfigure/MvcAutoConfiguration.java
@@ -3,10 +3,10 @@ package io.leangen.graphql.spqr.spring.autoconfigure;
 import graphql.GraphQL;
 import graphql.schema.GraphQLSchema;
 import io.leangen.graphql.spqr.spring.web.GraphQLController;
-import io.leangen.graphql.spqr.spring.web.servlet.GraphQLServletExecutor;
+import io.leangen.graphql.spqr.spring.web.mvc.GraphQLMvcExecutor;
 import io.leangen.graphql.spqr.spring.web.GuiController;
-import io.leangen.graphql.spqr.spring.web.servlet.DefaultGraphQLExecutor;
-import io.leangen.graphql.spqr.spring.web.servlet.DefaultGraphQLController;
+import io.leangen.graphql.spqr.spring.web.mvc.DefaultGraphQLExecutor;
+import io.leangen.graphql.spqr.spring.web.mvc.DefaultGraphQLController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -14,6 +14,8 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
+import static io.leangen.graphql.spqr.spring.autoconfigure.SpqrProperties.Http.Mvc.Executor.BLOCKING;
 
 @Configuration
 @ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.SERVLET)
@@ -25,22 +27,23 @@ public class MvcAutoConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    public ServletContextFactory globalContextFactory() {
+    public MvcContextFactory globalContextFactory() {
         return params -> new DefaultGlobalContext<>(params.getNativeRequest());
     }
 
     @Bean
     @ConditionalOnProperty(name = "graphql.spqr.http.enabled", havingValue = "true", matchIfMissing = true)
     @ConditionalOnMissingBean
-    public GraphQLServletExecutor graphQLExecutor(ServletContextFactory contextFactory) {
-        return new DefaultGraphQLExecutor(contextFactory, dataLoaderRegistryFactory);
+    public GraphQLMvcExecutor defaultExecutor(MvcContextFactory contextFactory, SpqrProperties config) {
+        DefaultGraphQLExecutor defaultExecutor = new DefaultGraphQLExecutor(contextFactory, dataLoaderRegistryFactory);
+        return config.getHttp().getMvc().getExecutor() == BLOCKING ? defaultExecutor.blocking() : defaultExecutor;
     }
 
     @Bean
     @ConditionalOnProperty(name = "graphql.spqr.http.enabled", havingValue = "true", matchIfMissing = true)
     @ConditionalOnMissingBean(GraphQLController.class)
     @ConditionalOnBean(GraphQLSchema.class)
-    public DefaultGraphQLController graphQLController(GraphQL graphQL, GraphQLServletExecutor executor) {
+    public DefaultGraphQLController graphQLController(GraphQL graphQL, GraphQLMvcExecutor executor) {
         return new DefaultGraphQLController(graphQL, executor);
     }