+++ graphql-spqr-spring-boot-autoconfigure/src/main/java/io/leangen/graphql/spqr/spring/autoconfigure/MvcContextFactory.java
@@ -2,5 +2,5 @@ package io.leangen.graphql.spqr.spring.autoconfigure;
 
 import org.springframework.web.context.request.NativeWebRequest;
 
-public interface ServletContextFactory extends ContextFactory<NativeWebRequest> {
+public interface MvcContextFactory extends ContextFactory<NativeWebRequest> {
 }