+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.6] - 2021-03-21
 ### Added
 - Support subscriptions over SSE [#97](https://github.com/leangen/graphql-spqr-spring-boot-starter/issues/97)
 
 ### Changed
+- [Breaking] Make async execution the default [#99](https://github.com/leangen/graphql-spqr-spring-boot-starter/issues/99)
 - Rewrite the WebSocket-based subscriptions implementation [#98](https://github.com/leangen/graphql-spqr-spring-boot-starter/issues/98)
+
+### Fixed
+- Bean registration fails when it encounters a JDK Proxy [#100](https://github.com/leangen/graphql-spqr-spring-boot-starter/issues/100)
 
 ## [0.0.5] - 2021-02-09
 ### Added